+++ src/ui.rs
@@ -15,6 +15,7 @@ use sdl2::video::{GLContext, SwapInterval};
 
 extern crate image;
 use crate::LUX_STEAM_DECK;
+use crate::LUX_STEAM_DECK_GAMING_MODE;
 
 const PROMPT_CONTROLLER_Y: &[u8] = include_bytes!("../res/prompts/Steam_Y.png");
 const PROMPT_CONTROLLER_A: &[u8] = include_bytes!("../res/prompts/Steam_A.png");
@@ -35,6 +36,11 @@ const PROMPT_KEYBOARD_SPACE: &[u8] = include_bytes!("../res/prompts/Space_Key_Da
 const PROMPT_KEYBOARD_ENTER: &[u8] = include_bytes!("../res/prompts/Enter_Key_Dark.png");
 const PROMPT_KEYBOARD_ESC: &[u8] = include_bytes!("../res/prompts/Esc_Key_Dark.png");
 const PROMPT_KEYBOARD_CTRL: &[u8] = include_bytes!("../res/prompts/Ctrl_Key_Dark.png");
+
+const PROMPT_CONTROLLER_STEAM_DECK_Y: &[u8] = include_bytes!("../res/prompts/SteamDeck_Y.png");
+const PROMPT_CONTROLLER_STEAM_DECK_A: &[u8] = include_bytes!("../res/prompts/SteamDeck_A.png");
+const PROMPT_CONTROLLER_STEAM_DECK_X: &[u8] = include_bytes!("../res/prompts/SteamDeck_X.png");
+const PROMPT_CONTROLLER_STEAM_DECK_B: &[u8] = include_bytes!("../res/prompts/SteamDeck_B.png");
 
 pub const DEFAULT_WINDOW_W: u32 = 600;
 pub const DEFAULT_WINDOW_H: u32 = 180;
@@ -63,6 +69,9 @@ impl Display for ControllerType {
             ControllerType::Switch => {
                 write!(f, "Switch")
             }
+            ControllerType::SteamDeck => {
+                write!(f, "SteamDeck")
+            }
         }
     }
 }
@@ -72,6 +81,7 @@ pub enum ControllerType {
     Xbox,
     DualShock,
     Switch,
+    SteamDeck,
 }
 
 pub struct EguiWindowInstance {
@@ -400,6 +410,7 @@ pub fn start_egui_window(
     window_flags |= sdl2::sys::SDL_WindowFlags::SDL_WINDOW_ALLOW_HIGHDPI as u32;
 
     let mut on_steam_deck = false;
+    let mut steam_deck_gaming_mode = true;
     match env::var(LUX_STEAM_DECK) {
         Ok(val) => {
             if val == "1" {
@@ -407,11 +418,26 @@ pub fn start_egui_window(
             }
         }
         Err(err) => {
-            debug!("SteamDeck env not found: {}", err);
+            debug!("LUX_STEAM_DECK env not found: {}", err);
         }
     }
 
-    info!("window is on display_index: {:?}", display_index);
+    match env::var(LUX_STEAM_DECK_GAMING_MODE) {
+        Ok(val) => {
+            if val != "1" {
+                steam_deck_gaming_mode = false;
+            }
+        }
+        Err(err) => {
+            debug!("LUX_STEAM_DECK_GAMING_MODE env not found: {}", err);
+            steam_deck_gaming_mode = false;
+        }
+    }
+
+    info!(
+        "window is on display_index: {:?} on_steam_deck: {} steam_deck_gaming_mode: {}",
+        display_index, on_steam_deck, steam_deck_gaming_mode
+    );
     match video_subsystem.display_dpi(display_index) {
         Ok(dpi) => {
             let mut using_dpi = dpi.0;
@@ -424,6 +450,7 @@ pub fn start_egui_window(
             }
 
             if on_steam_deck {
+                info!("halving dpi, since on steam deck");
                 using_dpi /= 2_f32;
             }
 
@@ -507,7 +534,9 @@ pub fn start_egui_window(
                     match game_controller_subsystem.name_for_index(id) {
                         Ok(name) => {
                             info!("controller name is {}", name);
-                            if name == "Steam Virtual Gamepad" && !on_steam_deck {
+                            if name == "Steam Virtual Gamepad"
+                                && (!on_steam_deck || !steam_deck_gaming_mode)
+                            {
                                 info!("ignorning steam virtual gamepad");
                                 ignore_controller = true;
                             }
@@ -546,6 +575,9 @@ pub fn start_egui_window(
                     } else if found_controller.name().contains("Pro") {
                         info!("controller assumed to be switch");
                         controller_type = ControllerType::Switch;
+                    } else if found_controller.name() == "Steam Virtual Gamepad" && on_steam_deck {
+                        info!("controller assumed to be steam deck");
+                        controller_type = ControllerType::SteamDeck;
                     } else {
                         info!("controller assumed to be xbox");
                     }
@@ -814,6 +846,8 @@ pub fn prompt_image_for_action(
                     image = PROMPT_CONTROLLER_DUALSHOCK_A;
                 } else if window_instance.controller_type == ControllerType::Switch {
                     image = PROMPT_CONTROLLER_SWITCH_A;
+                } else if window_instance.controller_type == ControllerType::SteamDeck {
+                    image = PROMPT_CONTROLLER_STEAM_DECK_A;
                 } else {
                     image = PROMPT_CONTROLLER_A;
                 }
@@ -827,6 +861,8 @@ pub fn prompt_image_for_action(
                     image = PROMPT_CONTROLLER_DUALSHOCK_B;
                 } else if window_instance.controller_type == ControllerType::Switch {
                     image = PROMPT_CONTROLLER_SWITCH_B;
+                } else if window_instance.controller_type == ControllerType::SteamDeck {
+                    image = PROMPT_CONTROLLER_STEAM_DECK_B;
                 } else {
                     image = PROMPT_CONTROLLER_B;
                 }
@@ -840,6 +876,8 @@ pub fn prompt_image_for_action(
                     image = PROMPT_CONTROLLER_DUALSHOCK_Y;
                 } else if window_instance.controller_type == ControllerType::Switch {
                     image = PROMPT_CONTROLLER_SWITCH_Y;
+                } else if window_instance.controller_type == ControllerType::SteamDeck {
+                    image = PROMPT_CONTROLLER_STEAM_DECK_Y;
                 } else {
                     image = PROMPT_CONTROLLER_Y;
                 }
@@ -853,6 +891,8 @@ pub fn prompt_image_for_action(
                     image = PROMPT_CONTROLLER_DUALSHOCK_X;
                 } else if window_instance.controller_type == ControllerType::Switch {
                     image = PROMPT_CONTROLLER_SWITCH_X;
+                } else if window_instance.controller_type == ControllerType::SteamDeck {
+                    image = PROMPT_CONTROLLER_STEAM_DECK_X;
                 } else {
                     image = PROMPT_CONTROLLER_X;
                 }